+++ nomis-ec-mode.el
@@ -111,8 +111,7 @@ Electric Clojure client and server code."
   "Face for Electric Clojure server code when using background color.")
 
 (defface nomis/ec-unparsable-face/using-background
-  `((t ,(list :box (list :color "Red"
-                         :line-width '(-1 . -1)))))
+  `((t ,(list :underline (list :color "Red"))))
   "Face for unparsable Electric Clojure code when using background color.
 This includes both bad syntax and parts of Clojure that we don't know about.")
 
@@ -131,8 +130,7 @@ This includes both bad syntax and parts of Clojure that we don't know about.")
   "Face for Electric Clojure server code when using underline.")
 
 (defface nomis/ec-unparsable-face/using-underline
-  `((t ,(list :box (list :color "Red"
-                         :line-width '(-1 . -1)))))
+  `((t ,(list :underline (list :color "Red"))))
   "Face for unparsable Electric Clojure code when using underline.
 This includes both bad syntax and parts of Clojure that we don't know about.")